requirements.txt
pymysql
cryptography
PyJWT
bcrypt
boto3
